Add docs EAR-1708
